Adjust margins, increase font size
The inner margin was much too thin given the binding, whereas the left and right margins looked excessive. This pushes the text more to the outside.

In addition, semiisopage uses slightly more vertical space, especially the footer seemed a bit high.

In order not to cram too much text on the page, this commit slightly increases the font size.

The end result is the same number of pages, but with a few extra lines on each page.