Increase top (and bottom) margin by 4mm
The latest proof copy lost 3mm at the top. This decreased the margin above the page number to 5mm, which is too much contrast with the margin between the number and the main text (1cm).

As a first attempt we maintain the symmetry (ratio: 1) with the bottom margin. This only slightly increases the page count. At least on screen the increased bottom margin does not seem excessive.